VHA_Backend/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
openai>=1.3.0
boto3>=1.34.0
redis>=5.0.1

# LangChain dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7

# Vector store and embeddings (CPU only to avoid GPU dependencies)
faiss-cpu>=1.7.4
spacy>=3.7.0

# Language detection
langdetect>=1.0.9

# Data processing
numpy>=1.24.3
pandas>=2.0.3

# Web framework (if using Flask/FastAPI)
flask>=2.3.3
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
flask-jwt-extended>=4.5.0
flasgger>=0.9.0

# Database
psycopg2-binary>=2.9.7

# Utilities
requests>=2.31.0
python-multipart>=0.0.6
Pillow>=10.0.1
pdfplumber>=0.9.0
PyPDF2>=3.0.0
pymupdf>=1.23.0
tiktoken>=0.5.0

# Authentication and security
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# Development dependencies
pytest>=7.4.3
black>=23.9.1
flake8>=6.1.0

# Additional dependencies that might be needed
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0
python-dateutil>=2.8.2
pypdf